+++ src/app/(site)/blog/[[...slug]]/avatar.tsx
